+++ crates/sncast/src/response/mod.rs
@@ -13,6 +13,7 @@ pub mod multicall;
 pub mod nonce;
 pub mod script;
 pub mod show_config;
+pub mod transaction;
 pub mod transformed_call;
 pub mod tx_status;
 pub mod ui;