+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Cast
+
+#### Added
+
+- `sncast get tx` command to get transaction details by hash.
+
 ## [0.58.1] - 2026-03-31
 
 ### Forge